requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
stripe==7.0.0
openai==1.54.0
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
Werkzeug==3.0.3
beautifulsoup4==4.12.3
APScheduler==3.10.4
